+++ apps/core-api/src/erp/purchasing/purchasing.module.ts
@@ -4,7 +4,7 @@ import { PurchasingService } from "./purchasing.service";
 
 @Module({
   controllers: [PurchasingController],
-  providers: [PurchasingService]
+  providers: [PurchasingService],
+  exports: [PurchasingService]
 })
 export class PurchasingModule {}
-